Refactor: Merge SystemAudioPermissionManager into SystemAudioManager
- Remove separate SystemAudioPermissionManager class
- Integrate permission handling directly into SystemAudioManager
- Follow same pattern as MicAudioManager for consistent architecture
- Add @Published permissionStatus property for UI binding
- Add permission methods: requestPermission(), checkPermissionStatus(), openSystemSettings()
- Update CaptionViewModel to use integrated permission system
- Eliminate architectural inconsistency between audio managers
- Reduce code complexity and improve maintainability

Benefits:
- Consistent permission handling across all audio managers
- Reduced coupling between CaptionViewModel and permission system
- Single responsibility: each audio manager handles its own permissions
- Cleaner API and reduced boilerplate code
