+++ Livcap/Archived/BufferManager.swift
