+++ Livcap/Archived/EnhancedVAD.swift
