+++ Livcap/Archived/LocalAgreementManager.swift
