+++ Livcap/Archived/TranscriptionStabilizationManager.swift
