+++ Livcap/Archived/WhisperCppTranscriber.swift
