+++ Livcap/Archived/WhisperLiveContinuousManager.swift
