+++ Livcap/Archived/WhisperLiveFrameBuffer.swift
