+++ Livcap/Archived/WhisperLiveFrameManager.swift
