+++ Livcap/Archived/WhisperLiveTranscriptionManager.swift
